+++ app/src/main/java/com/example/exchange/fragment/StartFragment.kt
@@ -1,0 +1,18 @@
+package com.example.exchange.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.exchange.R
+
+class StartFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_start, container, false)
+    }
+}
