+++ app/src/main/res/values/dimens.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <dimen name="margin_large">30dp</dimen>
+    <dimen name="text_size_answer">22sp</dimen>
 </resources>
