+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,6 @@
 <resources>
     <string name="app_name">Exchange</string>
-    <string name="conversion">Converter</string>
-    <string name="coin">Moedas</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="conversion">Conversor de Moedas</string>
+    <string name="coin">Lista de Moedas</string>
+    <string name="answer">O que você deseja visualizar?</string>
 </resources>
